Abide report start and end interval
With this patch, tagsum now parses the report start and end dates from
the parameters provided by timewarrior and uses them to calculate the
sum of time for each tag. Until now, when the start or stop date lay
outside the report interval, tagsum would output invalid times. This
was particularly notable when, e.g., running `timew tagsum today`, when
a multi-day recording existed.

Furthermore, this patch does the following:
* Improve type hints for function signatures
* Parse json data directly from stdin